make command line options logic reusable
