+++ src/generate_derived_key/generate_derived_key.cpp
@@ -14,338 +14,7 @@
 #include "parameters_eddsa.h"
 #include "parameters_ecdsa.h"
 #include "parameters_rsa.h"
-
-namespace {
-
-    /**
-     *  Utility function for reading a value from a stream, ensuring that a whole line is read
-     *
-     *  @param  stream     The stream to read from
-     *  @param  parser     Parser for the value from a stream
-     */
-    template <typename F,
-              typename = std::enable_if_t<std::is_nothrow_invocable_v<F, std::istringstream&>>>
-    void read_whole_line(std::istream &stream, F parser) noexcept
-    {
-        // read a line from the stream
-        std::string line;
-        std::getline(stream, line);
-
-        // read and parse the value
-        std::istringstream ss{line};
-        parser(ss);
-
-        // check whether the parse succeeded
-        if (!ss) {
-            // the parse failed, so report it and exit
-            std::cout << "Could not parse the string '" << line << "'" << std::endl;
-            exit(0);
-        }
-
-        // discard trailing whitespace
-        ss >> std::ws;
-
-        // check whether there is still content in the line
-        if (!ss.eof()) {
-            // not the whole line was consumed, so report it and exit
-            std::cout << "Could not parse the string '" << line << "': unused trailing characters" << std::endl;
-            exit(0);
-        }
-    }
-
-    /**
-     *  Which type of key should be generated?
-     */
-    enum class key_class {
-        eddsa,
-        ecdsa,
-        rsa2048,
-        rsa4096,
-        rsa8192,
-    };
-
-    /**
-     *  Reading a key_class from a stream parses the value
-     */
-    std::istream &operator>>(std::istream &stream, key_class &cl) noexcept
-    {
-        // read the value
-        read_whole_line(stream, [&cl](std::istream &s) noexcept {
-            // read a word from the stream
-            std::string word;
-            s >> word;
-
-            // attempt to parse the word
-            if (word == "eddsa") {
-                cl = key_class::eddsa;
-            } else if (word == "ecdsa") {
-                cl = key_class::ecdsa;
-            } else if (word == "rsa2048") {
-                cl = key_class::rsa2048;
-            } else if (word == "rsa4096") {
-                cl = key_class::rsa4096;
-            } else if (word == "rsa8192") {
-                std::cerr << "Warning: using an 8192-bit RSA key increases the chance of data leakage" << std::endl;
-                std::cerr << "Maximum recommended RSA key-size is 4096 bits" << std::endl;
-                cl = key_class::rsa8192;
-            } else {
-                std::cerr << "Unknown key type '" << word << "'" << std::endl;
-                // no parse, set the fail bit
-                s.setstate(std::ios_base::failbit);
-            }
-        });
-
-        // return the stream for chaining
-        return stream;
-    }
-
-    /**
-     *  Get a description of the key class
-     *
-     *  @param  type        The key class to get a description for
-     *  @return The description of that key class
-     */
-    constexpr boost::string_view key_class_description(key_class type) noexcept
-    {
-        switch (type) {
-            case key_class::eddsa:   return "EDDSA";
-            case key_class::ecdsa:   return "ECDSA";
-            case key_class::rsa2048: return "RSA 2048-bit";
-            case key_class::rsa4096: return "RSA 4096-bit";
-            case key_class::rsa8192: return "RSA 8192-bit";
-        }
-    }
-
-    /**
-     *  Wrapper around std::optional for reading from standard input with a prompt
-     *
-     *  If the type T is constructible with an std::string, it will be
-     *  constructed with a full line read from the input stream. Otherwise, the
-     *  value will be read directly from the stream using operator>>, and it
-     *  will be checked that it consumed the entire line.
-     */
-    template <typename T>
-    class opt_prompt : public std::optional<T> {
-    public:
-        // inherit the constructors
-        using std::optional<T>::optional;
-
-        /**
-         *  Read the value from standard input if there was none yet
-         *
-         *  @param  prompt     The prompt string to use when reading from stdin
-         */
-        void ensure_prompt(boost::string_view prompt) noexcept
-        {
-            // if we have a value already, nothing to do
-            if (this->has_value()) {
-                return;
-            }
-
-            // otherwise, prompt and read a value from input
-            std::cout << prompt << ": ";
-            std::cin >> *this;
-        }
-    };
-
-    /**
-     *  Reading an opt_prompt from a stream reads a line from the stream, then
-     *  parses the value from that line
-     */
-    template <typename T>
-    std::istream &operator>>(std::istream &stream, opt_prompt<T> &opt) noexcept
-    {
-        // read the value
-        read_whole_line(stream, [&opt](std::istream &stream) noexcept {
-            // if we can initialize the value with a string immediately, do so
-            if constexpr (std::is_constructible_v<T, const std::string &>) {
-                // read a line from the stream
-                std::string line;
-                std::getline(stream, line);
-
-                // initialize the value from that line
-                opt.emplace(line);
-            } else {
-                // read the value
-                T value;
-                stream >> value;
-
-                // if the parse succeeded, assign the value
-                if (stream) {
-                    opt = std::move(value);
-                }
-            }
-        });
-
-        // return the stream for chaining
-        return stream;
-    }
-
-    /**
-     *  A wrapper around an std::tm which enables reading it from standard input
-     */
-    class tm_wrapper : public std::tm {
-    public:
-        // inherit the constructors
-        using std::tm::tm;
-    };
-
-    /**
-     *  Parse an std::tm from a stream
-     */
-    std::istream &operator>>(std::istream &stream, tm_wrapper &tm) noexcept
-    {
-        // read the value
-        read_whole_line(stream, [&tm](std::istream &s) noexcept {
-            // explicitly construct an instance to zero-initialize the structure
-            tm = tm_wrapper{};
-            s >> std::get_time(&tm, "%Y-%m-%d %H:%M:%S");
-        });
-
-        // return the stream for chaining
-        return stream;
-    }
-
-    /**
-     *  Parse a boolean from a stream
-     */
-    std::istream &operator>>(std::istream &stream, bool &result) noexcept
-    {
-        // the inputs to be used for true and false
-        constexpr boost::string_view yes { "yes" };
-        constexpr boost::string_view no  { "no"  };
-
-        // read the user input
-        std::string input { "invalid" };
-        std::getline(stream, input);
-
-        // convert the input to lowercase
-        std::transform(input.begin(), input.end(), input.begin(), [](unsigned char c) {
-            // convert the character to lowercase
-            return std::tolower(c);
-        });
-
-        // did the user select yes (default)
-        if (input.empty() || yes.starts_with(input)) {
-            // the user either typed 'yes' or did not select anything
-            result = true;
-        } else if (no.starts_with(input)) {
-            // the user typed no
-            result = false;
-        } else {
-            // no parse, set the fail bit
-            stream.setstate(std::ios_base::failbit);
-        }
-
-        // return the stream for chaining
-        return stream;
-    }
-
-    /**
-     *  The parsed program options.
-     */
-    struct Options {
-        // the key output file
-        opt_prompt<std::string> output_file;
-
-        // the type of the key
-        opt_prompt<key_class>   type;
-
-        // the user id
-        opt_prompt<std::string> user_name;
-        opt_prompt<std::string> user_email;
-
-        // the start and end of signature validity
-        opt_prompt<tm_wrapper>  signature_creation;
-        opt_prompt<tm_wrapper>  signature_expiration;
-
-        // meta-information for the key derivation
-        opt_prompt<std::string> kdf_context;
-        opt_prompt<tm_wrapper>  key_creation;
-
-        bool                    debug_dump_keys;
-    };
-
-    /**
-     *  Parse options from the program command-line arguments.
-     *
-     *  @param  argc    The 'argc' parameter to main()
-     *  @param  argv    The 'argv' parameter to main()
-     */
-    Options parse_options(int argc, const char **argv)
-    {
-        namespace po = boost::program_options;
-
-        // the options structure that we will configure here
-        Options options;
-
-        // description of the options for the boost option parser
-        // Set the line length to 100 for slightly wider option descriptions
-        po::options_description optdesc(100);
-        optdesc.add_options()
-            ("help,h",                                                                            "Produce help message")
-            ("output-file,o",  po::value<opt_prompt<std::string>>(&options.output_file),          "Output file")
-            ("key-type,t",     po::value<opt_prompt<key_class>>  (&options.type),                 "Type of the generated key (eddsa/ecdsa/rsa{2048,4096,8192})")
-            ("name,n",         po::value<opt_prompt<std::string>>(&options.user_name),            "Your name (firstname lastname)")
-            ("email,e",        po::value<opt_prompt<std::string>>(&options.user_email),           "Your email address")
-            ("sigtime,s",      po::value<opt_prompt<tm_wrapper>> (&options.signature_creation),   "Signature creation time in UTC (YYYY-MM-DD HH:MM:SS)")
-            ("sigexpiry,x",    po::value<opt_prompt<tm_wrapper>> (&options.signature_expiration), "Signature expiration time in UTC (YYYY-MM-DD HH:MM:SS)")
-            ("kdf-context,k",  po::value<opt_prompt<std::string>>(&options.kdf_context),          "Key derivation context (8 bytes)")
-            ("key-creation,c", po::value<opt_prompt<tm_wrapper>> (&options.key_creation),         "Key creation time in UTC (YYYY-MM-DD HH:MM:SS)")
-            ("debug-dump-secret-and-public-keys", po::bool_switch(&options.debug_dump_keys),      "Dump generated key parameters; WARNING: sensitive data!");
-
-        // run the option parser
-        po::variables_map vm;
-        po::store(
-            po::command_line_parser{argc, argv}.options(optdesc).run(),
-            vm
-        );
-        po::notify(vm);
-
-        // check for the help flag
-        if (vm.count("help")) {
-            // output generic info
-            std::cout << "Usage: " << argv[0] << " [options]" << std::endl;
-            std::cout << std::endl;
-            std::cout << "This program will deterministically generate a PGP key based on user-provided entropy." << std::endl;
-            std::cout << "The program will prompt for certain required input, after prompting to obtain information from" << std::endl;
-            std::cout << "command-line options." << std::endl;
-            std::cout << std::endl;
-            std::cout << "Note that the generated signatures may not be deterministic, since making cryptographic signatures" << std::endl;
-            std::cout << "is in general a non-deterministic process. The key, however, is deterministic." << std::endl;
-            std::cout << std::endl;
-            std::cout << "This program is a work-in-progress, and is not adequately documented yet. Proceed with caution." << std::endl;
-            std::cout << std::endl;
-
-            // output the generated help text for the options
-            std::cout << optdesc << std::endl;
-
-            // exit, since the user just requested help
-            exit(0);
-        }
-
-        // ensure that all the options are initialized by possibly reading some from standard input
-        options.output_file         .ensure_prompt("Output file");
-        options.type                .ensure_prompt("Type of the generated key (eddsa/ecdsa/rsa{2048,4096,8192})");
-        options.user_name           .ensure_prompt("Your name (firstname lastname)");
-        options.user_email          .ensure_prompt("Your email address");
-        options.signature_creation  .ensure_prompt("Signature creation time in UTC (YYYY-MM-DD HH:MM:SS)");
-        options.signature_expiration.ensure_prompt("Signature expiration time in UTC (YYYY-MM-DD HH:MM:SS)");
-        options.kdf_context         .ensure_prompt("Key derivation context (8 bytes)");
-        options.key_creation        .ensure_prompt("Key creation time in UTC (YYYY-MM-DD HH:MM:SS)");
-
-        // check that the KDF context is the right size
-        if (options.kdf_context->size() != 8) {
-            // alert the user to the invalid input and exit
-            std::cerr << "Key derivation context should be exactly 8 bytes, but is " << options.kdf_context->size() << " bytes" << std::endl;
-            exit(1);
-        }
-
-        // return the created options struct
-        return options;
-    }
-
-}
+#include "options.h"
 
 /**
  *  Main function
@@ -365,10 +34,10 @@ int main(int argc, const char **argv)
         }
 
         // parse the command-line arguments
-        Options options = parse_options(argc, argv);
+        key_generation::options options{ argc, argv };
 
         // inform the user about the settings in the command-line arguments
-        std::cout << "Using key type " << key_class_description(*options.type) << std::endl;
+        std::cout << "Using key type " << util::key_class_description(*options.type) << std::endl;
         std::cout << "Writing key to file '" << *options.output_file << "'" << std::endl;
 
         // the master key for generation
@@ -433,11 +102,11 @@ int main(int argc, const char **argv)
         // select the function with which to generate the packets
         std::function<std::vector<pgp::packet>(const master_key&, std::string, uint32_t, uint32_t, uint32_t, boost::string_view, bool)> generation_function;
         switch (*options.type) {
-            case key_class::eddsa: generation_function = generate_key<parameters::eddsa>; break;
-            case key_class::ecdsa: generation_function = generate_key<parameters::ecdsa>; break;
-            case key_class::rsa2048: generation_function = generate_key<parameters::rsa<2048>>; break;
-            case key_class::rsa4096: generation_function = generate_key<parameters::rsa<4096>>; break;
-            case key_class::rsa8192: generation_function = generate_key<parameters::rsa<8192>>; break;
+            case util::key_class::eddsa: generation_function = generate_key<parameters::eddsa>; break;
+            case util::key_class::ecdsa: generation_function = generate_key<parameters::ecdsa>; break;
+            case util::key_class::rsa2048: generation_function = generate_key<parameters::rsa<2048>>; break;
+            case util::key_class::rsa4096: generation_function = generate_key<parameters::rsa<4096>>; break;
+            case util::key_class::rsa8192: generation_function = generate_key<parameters::rsa<8192>>; break;
         }
 
         // generate the packets
